feat: reorganize bot files
