+++ Dockerfile
@@ -19,4 +19,4 @@ ARG version=local
 ENV IMAGE_TAG=$version
 
 EXPOSE 8080
-CMD ["python3", "main.py"]
+CMD ["python3", "-m", "chouette"]
