+++ chouette/commands/misc.py
@@ -6,11 +6,11 @@ from typing import TYPE_CHECKING
 import discord
 from discord import app_commands
 
-from utils.github_api import get_last_update
-from utils.latex_render import latex_render
+from chouette.utils.github_api import get_last_update
+from chouette.utils.latex_render import latex_render
 
 if TYPE_CHECKING:
-    from bot import ChouetteBot
+    from chouette.bot import ChouetteBot
 
 
 # Make a LaTeX command
