+++ chouette/utils/version.py
